feat: separate component from page
